Refactor imports in `ProcessStopEvent` and `ProcessExceptionEvent` to use explicit paths for `WorkEvent`.
